moving classes from layouts.layouts
