+++ graspologic/layouts/__init__.py
@@ -1,6 +1,8 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
+from graspologic.layouts.classes import NodePosition
+from graspologic.layouts.visualize import save_graph, show_graph
 from graspologic.layouts import layouts, render_only, tsne, nooverlap
 from graspologic.layouts.layout_from_edges import (
     layout_node2vec_tsne_from_file,
@@ -11,7 +13,9 @@ from graspologic.layouts.layout_from_edges import (
 )
 
 __all__ = [
-    "layouts",
+    "NodePositions",
+    "save_graph",
+    "show_graph",
     "_helpers",
     "render_only",
     "tsne",
