--- graspologic/layouts/layouts/__init__.py
@@ -1,13 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT license.
-
-
-from graspologic.layouts.layouts.classes import NodePosition
-
-from graspologic.layouts.layouts.visualize import save_graph, show_graph
-
-__all__ = [
-    "NodePosition",
-    "save_graph",
-    "show_graph",
-]
